adf_hd refactoring: code for generic device and floppy device moved to separated modules.
