+++ src/adf_dev_flop.h
@@ -1,0 +1,14 @@
+
+#ifndef __ADF_DEV_FLOP_H__
+#define __ADF_DEV_FLOP_H__
+
+#include "adf_defs.h"
+#include "adf_str.h"
+#include "prefix.h"
+
+RETCODE adfMountFlop ( struct Device * dev );
+
+PREFIX RETCODE adfCreateFlop ( struct Device * dev,
+                               char *          volName,
+                               int             volType );
+#endif
